refactor(setup): drive tool integration from a registry
Replace the per-tool switch statements (detect, configure, parse,
display name) with a single registry plus one file per tool, and
collapse the duplicated routing-guidance blobs into one shared source
(guidanceMarkdown). Behavior-preserving: adding a tool is now a new
file with a detect/configure pair plus one registry line, with no
switch sites to update. The CLI help and --tools list now include
OpenCode.
