+++ internal/setup/opencode.go
@@ -3,31 +3,36 @@ package setup
 import (
 	"fmt"
 	"os"
+	"os/exec"
 	"path/filepath"
 	"strings"
 )
 
-const opencodeAgentsMD = `<!-- sense:start -->
-## Sense — codebase understanding
+// detectOpencode looks for evidence that OpenCode is installed.
+func detectOpencode() DetectResult {
+	r := DetectResult{Tool: ToolOpencode}
+	if os.Getenv("OPENCODE") != "" {
+		r.Found = true
+		r.Evidence = "OPENCODE env var set"
+		return r
+	}
+	if _, err := exec.LookPath("opencode"); err == nil {
+		r.Found = true
+		r.Evidence = "opencode on PATH"
+		return r
+	}
+	if home, err := os.UserHomeDir(); err == nil {
+		if _, err := os.Stat(filepath.Join(home, ".config", "opencode")); err == nil {
+			r.Found = true
+			r.Evidence = "~/.config/opencode/ directory"
+			return r
+		}
+	}
+	return r
+}
 
-This project has a Sense index. Sense gives you structural understanding of the codebase — symbols, relationships, patterns — without reading dozens of files.
-
-**Use Sense MCP tools for ALL codebase understanding:**
-
-| Question | Tool |
-|---|---|
-| Who calls X? What does X call? | sense_graph |
-| Find code related to a concept | sense_search |
-| What breaks if I change X? | sense_blast |
-| What patterns does this project follow? | sense_conventions |
-| Index health, what's indexed | sense_status |
-
-**When NOT to use Sense** (use grep instead):
-- Exact text/string search (regex, log messages, string literals)
-- Reading file contents → use your file reading tool
-- Editing code → Sense is read-only
-<!-- sense:end -->`
-
+// configureOpencode writes OpenCode's project-local MCP config, the AGENTS.md
+// guidance, and the per-skill SKILL.md tree OpenCode expects.
 func configureOpencode(root string) (*ToolResult, error) {
 	tr := &ToolResult{Tool: ToolOpencode}
 
@@ -88,7 +93,7 @@ func writeOpencodeJSON(root string) (bool, error) {
 
 // writeOpencodeAgentsMD creates or updates the Sense section in AGENTS.md.
 func writeOpencodeAgentsMD(root string) (bool, error) {
-	return writeMarkerFile(filepath.Join(root, "AGENTS.md"), opencodeAgentsMD)
+	return writeMarkerFile(filepath.Join(root, "AGENTS.md"), guidanceMarkdown)
 }
 
 // writeOpencodeSkills creates skill files in .opencode/skills/.
